pyproject.toml
[tool.poetry]
name = "letmecook"
version = "0.1.0"
description = ""
authors = ["suedaltprmk <suedaltprmk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pymongo = "^4.7.1"
uvicorn = "^0.29.0"
pytest = "^8.2.0"
httpx = "^0.27.0"
faker = "^25.0.1"
fastapi-jwt-auth = "^0.5.0"
requests = "^2.31.0"
aliyun-python-sdk-airec = "1.1.3"
oss2 = "^2.18.5"
pillow = "^10.3.0"
pip = "^24.0"
install = "^1.3.5"
aliyun-python-sdk-core = "^2.15.1"
transformers = "^4.41.1"
torch = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
